+++ tidy3d/web/api/autograd/autograd.py
@@ -13,7 +13,7 @@ from tidy3d.components.autograd.derivative_utils import DerivativeInfo
 
 from ..asynchronous import DEFAULT_DATA_DIR
 from ..asynchronous import run_async as run_async_webapi
-from ..container import BatchData
+from ..container import BatchData, Job
 from ..tidy3d_stub import SimulationDataType, SimulationType
 from ..webapi import run as run_webapi
 from .utils import get_derivative_maps, split_data_list, split_list
@@ -21,6 +21,8 @@ from .utils import get_derivative_maps, split_data_list, split_list
 # keys for data into auxiliary dictionary
 AUX_KEY_SIM_DATA_ORIGINAL = "sim_data"
 AUX_KEY_SIM_DATA_FWD = "sim_data_fwd_adjoint"
+AUX_KEY_FWD_TASK_ID = "task_id_fwd"
+
 ISSUE_URL = (
     "https://github.com/flexcompute/tidy3d/issues/new?"
     "assignees=tylerflex&labels=adjoint&projects=&template=autograd_bug.md"
@@ -28,6 +30,9 @@ ISSUE_URL = (
 URL_LINK = f"[blue underline][link={ISSUE_URL}]'{ISSUE_URL}'[/link][/blue underline]"
 
 MAX_NUM_TRACED_STRUCTURES = 500
+
+# default value for whether to do local gradient calculation (True) or server side (False)
+LOCAL_GRADIENT = True
 
 
 def warn_autograd(fn_name: str, exc: Exception) -> str:
@@ -93,6 +98,7 @@ def run(
     worker_group: str = None,
     simulation_type: str = "tidy3d",
     parent_tasks: list[str] = None,
+    local_gradient: bool = LOCAL_GRADIENT,
 ) -> SimulationDataType:
     """
     Submits a :class:`.Simulation` to server, starts running, monitors progress, downloads,
@@ -123,6 +129,9 @@ def run(
         target solver version.
     worker_group: str = None
         worker group
+    local_gradient: bool = False
+        Whether to perform gradient calculation locally, requiring more downloads but potentially
+        more stable with experimental features.
 
     Returns
     -------
@@ -184,6 +193,7 @@ def run(
                 worker_group=worker_group,
                 simulation_type="tidy3d_autograd",
                 parent_tasks=parent_tasks,
+                local_gradient=local_gradient,
             )
         except Exception as exc:
             warn_autograd("web.run()", exc=exc)
@@ -281,7 +291,9 @@ def run_async(
 """ User-facing ``run`` and `run_async`` functions, compatible with ``autograd`` """
 
 
-def _run(simulation: td.Simulation, task_name: str, **run_kwargs) -> td.SimulationData:
+def _run(
+    simulation: td.Simulation, task_name: str, local_gradient: bool = LOCAL_GRADIENT, **run_kwargs
+) -> td.SimulationData:
     """User-facing ``web.run`` function, compatible with ``autograd`` differentiation."""
 
     traced_fields_sim = setup_run(simulation=simulation)
@@ -294,7 +306,8 @@ def _run(simulation: td.Simulation, task_name: str, **run_kwargs) -> td.Simulati
             "to the 'Simulation'. If this is unexpected, double check your objective function "
             "pre-processing. Running regular tidy3d simulation."
         )
-        return _run_tidy3d(simulation, task_name=task_name, **run_kwargs)
+        data, _ = _run_tidy3d(simulation, task_name=task_name, **run_kwargs)
+        return data
 
     # will store the SimulationData for original and forward so we can access them later
     aux_data = {}
@@ -305,6 +318,7 @@ def _run(simulation: td.Simulation, task_name: str, **run_kwargs) -> td.Simulati
         sim_original=simulation.to_static(),
         task_name=task_name,
         aux_data=aux_data,
+        local_gradient=local_gradient,
         **run_kwargs,
     )
 
@@ -374,16 +388,30 @@ def _run_primitive(
     sim_original: td.Simulation,
     task_name: str,
     aux_data: dict,
+    local_gradient: bool,
     **run_kwargs,
 ) -> AutogradFieldMap:
     """Autograd-traced 'run()' function: runs simulation, strips tracer data, caches fwd data."""
 
     td.log.info("running primitive '_run_primitive()'")
     sim_combined = setup_fwd(sim_fields=sim_fields, sim_original=sim_original)
-    sim_data_combined = _run_tidy3d(sim_combined, task_name=task_name, **run_kwargs)
-    return postprocess_fwd(
-        sim_data_combined=sim_data_combined, sim_original=sim_original, aux_data=aux_data
+
+    if not local_gradient:
+        run_kwargs["simulation_type"] = "autograd_fwd"
+
+    sim_data_combined, task_id_fwd = _run_tidy3d(sim_combined, task_name=task_name, **run_kwargs)
+
+    # TODO: put this in postprocess?
+    aux_data[AUX_KEY_FWD_TASK_ID] = task_id_fwd
+
+    # TODO: make sure this function works if the combined sim data has no grad monitors
+    field_map = postprocess_fwd(
+        sim_data_combined=sim_data_combined,
+        sim_original=sim_original,
+        aux_data=aux_data,
+        local_gradient=local_gradient,
     )
+    return field_map
 
 
 @primitive
@@ -409,7 +437,10 @@ def _run_async_primitive(
         sim_original = sims_original[task_name]
         aux_data = aux_data_dict[task_name]
         field_map_fwd_dict[task_name] = postprocess_fwd(
-            sim_data_combined=sim_data_combined, sim_original=sim_original, aux_data=aux_data
+            sim_data_combined=sim_data_combined,
+            sim_original=sim_original,
+            aux_data=aux_data,
+            local_gradient=False,  # TODO: fix once we start supporting
         )
 
     return field_map_fwd_dict
@@ -423,7 +454,10 @@ def setup_fwd(sim_fields: AutogradFieldMap, sim_original: td.Simulation) -> td.S
 
 
 def postprocess_fwd(
-    sim_data_combined: td.SimulationData, sim_original: td.Simulation, aux_data: dict
+    sim_data_combined: td.SimulationData,
+    sim_original: td.Simulation,
+    aux_data: dict,
+    local_gradient: bool,
 ) -> AutogradFieldMap:
     """Postprocess the combined simulation data into an Autograd field map."""
 
@@ -441,20 +475,20 @@ def postprocess_fwd(
     sim_data_original = sim_data_combined.updated_copy(
         simulation=sim_original, data=data_original, deep=False
     )
-
-    # construct the 'forward' simulation and its data, which is only used for for gradient calc.
-    sim_fwd = sim_combined.updated_copy(monitors=monitors_fwd)
-    sim_data_fwd = sim_data_combined.updated_copy(
-        simulation=sim_fwd,
-        data=data_fwd,
-        deep=False,
-    )
-
-    # cache these two SimulationData objects for later (note: the Simulations are already inside)
     aux_data[AUX_KEY_SIM_DATA_ORIGINAL] = sim_data_original
-    aux_data[AUX_KEY_SIM_DATA_FWD] = sim_data_fwd
 
-    # strip out the tracer AutogradFieldMap for the .data from the original sim
+    if local_gradient:
+        # construct the 'forward' simulation and its data, which is only used for for gradient calc.
+        sim_fwd = sim_combined.updated_copy(monitors=monitors_fwd)
+        sim_data_fwd = sim_data_combined.updated_copy(
+            simulation=sim_fwd,
+            data=data_fwd,
+            deep=False,
+        )
+        aux_data[AUX_KEY_SIM_DATA_FWD] = sim_data_fwd
+
+        # strip out the tracer AutogradFieldMap for the .data from the original sim
+
     data_traced = sim_data_original.strip_traced_fields(
         include_untraced_data_arrays=True, starting_path=("data",)
     )
@@ -472,13 +506,14 @@ def _run_bwd(
     sim_original: td.Simulation,
     task_name: str,
     aux_data: dict,
+    local_gradient: bool,
     **run_kwargs,
 ) -> typing.Callable[[AutogradFieldMap], AutogradFieldMap]:
     """VJP-maker for ``_run_primitive()``. Constructs and runs adjoint simulation, computes grad."""
 
     # get the fwd epsilon and field data from the cached aux_data
     sim_data_orig = aux_data[AUX_KEY_SIM_DATA_ORIGINAL]
-    sim_data_fwd = aux_data[AUX_KEY_SIM_DATA_FWD]
+    sim_data_fwd = aux_data.get(AUX_KEY_SIM_DATA_FWD)
 
     td.log.info("constructing custom vjp function for backwards pass.")
 
@@ -509,7 +544,20 @@ def _run_bwd(
 
         # run adjoint simulation
         task_name_adj = str(task_name) + "_adjoint"
-        sim_data_adj = _run_tidy3d(sim_adj, task_name=task_name_adj, **run_kwargs)
+
+        if not local_gradient:
+            run_kwargs["simulation_type"] = "autograd_bwd"
+            task_id_fwd = aux_data[AUX_KEY_FWD_TASK_ID]
+            run_kwargs["parent_tasks"] = [task_id_fwd]
+            # TODO: run the adjoint special web API? get the field map
+            # job = Job(...)
+            # job.upload()
+            # job.start()
+            # job.monitor()
+            # traced_fields_vjp = download_sim_vjp_autograd(job.task_id)
+            # return traced_fields_vjp
+
+        sim_data_adj, _ = _run_tidy3d(sim_adj, task_name=task_name_adj, **run_kwargs)
 
         return postprocess_adj(
             sim_data_adj=sim_data_adj,
@@ -685,12 +733,17 @@ defvjp(_run_async_primitive, _run_async_bwd, argnums=[0])
 """ The fundamental Tidy3D run and run_async functions used above. """
 
 
-def _run_tidy3d(simulation: td.Simulation, task_name: str, **run_kwargs) -> td.SimulationData:
+def _run_tidy3d(
+    simulation: td.Simulation, task_name: str, **run_kwargs
+) -> (td.SimulationData, str):
     """Run a simulation without any tracers using regular web.run()."""
     td.log.info("running regular simulation with '_run_tidy3d()'")
     # TODO: set task_type to "tidy3d adjoint autograd?"
-    data = run_webapi(simulation, task_name=task_name, **run_kwargs)
-    return data
+
+    run_kwargs = {k: v for k, v in run_kwargs.items() if k in Job._upload_fields}
+    job = Job(simulation=simulation, task_name=task_name, **run_kwargs)
+    data = job.run()
+    return data, job.task_id
 
 
 def _run_async_tidy3d(simulations: dict[str, td.Simulation], **run_async_kwargs) -> BatchData: